feat: Next.js 15 (App Router) + TypeScriptに移行
- Create React AppからNext.js 15 (App Router)に完全移行
- React 19 + TypeScript 5.3に更新
- styled-componentsのSSR対応を追加
- react-router-domからnext/linkへ移行
- ディレクトリ構造をApp Router形式に再編成
  - app/: ページルーティング
  - components/: 共通コンポーネント
  - lib/: ユーティリティと定数
