+++ components/Footer/styled.tsx
@@ -1,0 +1,34 @@
+'use client';
+
+import styled from 'styled-components';
+
+const StyledFooter = styled.div`
+  footer {
+    background-color: #333;
+    color: #fff;
+    padding: 20px 0;
+    text-align: center;
+    font-size: 14px;
+  }
+
+  footer a {
+    color: #fff;
+    text-decoration: none;
+    margin: 0 10px;
+  }
+
+  footer p {
+    margin-top: 10px;
+  }
+
+  footer ul {
+    list-style-type: none;
+    padding: 0;
+  }
+
+  footer ul li {
+    display: inline;
+  }
+`;
+
+export default StyledFooter;
