+++ lib/constants/slotSettings.ts
