+++ next.config.ts
@@ -1,0 +1,9 @@
+import type { NextConfig } from 'next';
+
+const nextConfig: NextConfig = {
+  compiler: {
+    styledComponents: true,
+  },
+};
+
+export default nextConfig;
