--- public/data.json
@@ -1,15 +1,0 @@
-{
-  "about": {
-    "title": "開発者について",
-    "content": "はじめまして、しょぺこと申します。現役の兼業エナ専です。本業はエンジニアをしています。簡単に私の人生について話させて頂きます。私は、高校卒業後現役で歯学部に入学しまし、在学中にスロットに出会いました。友達に連れられて行って初めて打ったのは”バジリスク絆”でした。５千円勝ちました。少額ではあったものの”勝った快感”が忘れられず、次の日一人で打ちに行き、”バジリスク２”で８万円ほど勝ちました。それからどっぷりハマり、平打ちをしばらく続けましたがビギナーズラックで勝ちつづけ１ヶ月で２０万円以上勝ったのを覚えています。それまでバジリスクシリーズを打ち続けていましたが、”寺井一択”さんの”ハーデス”の動画を見て、ハーデスを打ちに行ったところ、６万円負けました。初めて負け、しかも６万円という大金を一瞬で失ってスロットが怖くなり、一旦スロットを引退しました。しかし、数ヶ月後親戚とうちに行くことになり、また負けました。この時から、スロットでどうやって勝つことができるのか調べ始め、”期待値稼働”の存在を知ります。学生の頃から数学が好きだった私は、すぐにスロットで勝つことができることを理解し、実践し、稼げるようになりました。それから、以前から辞めたいと思っていた大学を中退し、専業になります。専業は2年ほど続けましたが、ピンでは稼げる額に限界があることと、プログラミングに興味があったことから、再度大学に入学しました。なので現在は、大学で数学を学びつづ、独学でプログラミングをしています。"
-  },
-  "profile": {
-    "title": "プロフィール",
-    "content": {
-      "name": "しょぺこ",
-      "favorite": "沖ドキ、北斗、ラブ嬢",
-      "carrier": "専業2年、兼業7年",
-      "hobby": "ネトゲ、麻雀、アニメ"
-    }
-  }
-}
