--- src/App.css
@@ -1,4 +1,0 @@
-.App {
-  text-align: center;
-}
-
