--- src/pages/NotFound.tsx
@@ -1,15 +1,0 @@
-import React from "react";
-import Footer from "../component/Footer";
-import Header from "../component/Header";
-
-function NotFound() {
-  return (
-    <>
-      <Header />
-      <div>404: Page Not Found</div>
-      <Footer />
-    </>
-  );
-}
-
-export default NotFound;
